+++ sea-orm-macros/src/derives/entity_model.rs
@@ -63,6 +63,8 @@ pub fn expand_derive_entity_model(data: Data, attrs: Vec<Attribute>) -> syn::Res
     // generate Column enum and it's ColumnTrait impl
     let mut columns_enum: Punctuated<_, Comma> = Punctuated::new();
     let mut columns_trait: Punctuated<_, Comma> = Punctuated::new();
+    let mut columns_select_as: Punctuated<_, Comma> = Punctuated::new();
+    let mut columns_save_as: Punctuated<_, Comma> = Punctuated::new();
     let mut primary_keys: Punctuated<_, Comma> = Punctuated::new();
     let mut primary_key_types: Punctuated<_, Comma> = Punctuated::new();
     let mut auto_increment = true;
@@ -90,6 +92,8 @@ pub fn expand_derive_entity_model(data: Data, attrs: Vec<Attribute>) -> syn::Res
                     let mut nullable = false;
                     let mut default_value = None;
                     let mut default_expr = None;
+                    let mut select_as = None;
+                    let mut save_as = None;
                     let mut indexed = false;
                     let mut ignore = false;
                     let mut unique = false;
@@ -169,6 +173,24 @@ pub fn expand_derive_entity_model(data: Data, attrs: Vec<Attribute>) -> syn::Res
                                                         format!("Invalid enum_name {:?}", nv.lit),
                                                     ));
                                                 }
+                                            } else if name == "select_as" {
+                                                if let Lit::Str(litstr) = &nv.lit {
+                                                    select_as = Some(litstr.value());
+                                                } else {
+                                                    return Err(Error::new(
+                                                        field.span(),
+                                                        format!("Invalid select_as {:?}", nv.lit),
+                                                    ));
+                                                }
+                                            } else if name == "save_as" {
+                                                if let Lit::Str(litstr) = &nv.lit {
+                                                    save_as = Some(litstr.value());
+                                                } else {
+                                                    return Err(Error::new(
+                                                        field.span(),
+                                                        format!("Invalid save_as {:?}", nv.lit),
+                                                    ));
+                                                }
                                             }
                                         }
                                     }
@@ -221,6 +243,23 @@ pub fn expand_derive_entity_model(data: Data, attrs: Vec<Attribute>) -> syn::Res
                         primary_keys.push(quote! {
                             #variant_attrs
                             #field_name
+                        });
+                    }
+
+                    if let Some(select_as) = select_as {
+                        columns_select_as.push(quote! {
+                            Self::#field_name => sea_orm::sea_query::SimpleExpr::cast_as(
+                                Into::<sea_orm::sea_query::SimpleExpr>::into(expr),
+                                sea_orm::sea_query::Alias::new(&#select_as),
+                            ),
+                        });
+                    }
+                    if let Some(save_as) = save_as {
+                        columns_save_as.push(quote! {
+                            Self::#field_name => sea_orm::sea_query::SimpleExpr::cast_as(
+                                Into::<sea_orm::sea_query::SimpleExpr>::into(val),
+                                sea_orm::sea_query::Alias::new(&#save_as),
+                            ),
                         });
                     }
 
@@ -348,6 +387,20 @@ pub fn expand_derive_entity_model(data: Data, attrs: Vec<Attribute>) -> syn::Res
             fn def(&self) -> sea_orm::prelude::ColumnDef {
                 match self {
                     #columns_trait
+                }
+            }
+
+            fn select_as(&self, expr: sea_orm::sea_query::Expr) -> sea_orm::sea_query::SimpleExpr {
+                match self {
+                    #columns_select_as
+                    _ => sea_orm::prelude::ColumnTrait::select_enum_as(self, expr),
+                }
+            }
+
+            fn save_as(&self, val: sea_orm::sea_query::Expr) -> sea_orm::sea_query::SimpleExpr {
+                match self {
+                    #columns_save_as
+                    _ => sea_orm::prelude::ColumnTrait::save_enum_as(self, val),
                 }
             }
         }