+++ src/query/combine.rs
@@ -1,6 +1,5 @@
 use crate::{
-    cast_enum_as_text, ColumnTrait, EntityTrait, IdenStatic, Iterable, QueryTrait, Select,
-    SelectTwo, SelectTwoMany,
+    ColumnTrait, EntityTrait, IdenStatic, Iterable, QueryTrait, Select, SelectTwo, SelectTwoMany,
 };
 use core::marker::PhantomData;
 pub use sea_query::JoinType;
@@ -149,7 +148,7 @@ where
     for col in <F::Column as Iterable>::iter() {
         let alias = format!("{}{}", SelectB.as_str(), col.as_str());
         selector.query().expr(SelectExpr {
-            expr: cast_enum_as_text(col.into_expr(), &col),
+            expr: col.select_as(col.into_expr()),
             alias: Some(SeaRc::new(Alias::new(&alias))),
             window: None,
         });