+++ tests/common/features/schema.rs
@@ -335,6 +335,12 @@ pub async fn create_json_struct_table(db: &DbConn) -> Result<ExecResult, DbErr> 
 }
 
 pub async fn create_collection_table(db: &DbConn) -> Result<ExecResult, DbErr> {
+    db.execute(sea_orm::Statement::from_string(
+        db.get_database_backend(),
+        "CREATE EXTENSION IF NOT EXISTS citext".into(),
+    ))
+    .await?;
+
     let stmt = sea_query::Table::create()
         .table(collection::Entity)
         .col(
@@ -343,6 +349,11 @@ pub async fn create_collection_table(db: &DbConn) -> Result<ExecResult, DbErr> {
                 .not_null()
                 .auto_increment()
                 .primary_key(),
+        )
+        .col(
+            ColumnDef::new(collection::Column::Name)
+                .custom(Alias::new("citext"))
+                .not_null(),
         )
         .col(
             ColumnDef::new(collection::Column::Integers)